Add files via upload
This code uses K-means clustering algorithm to group mall customers based on their annual income and spending score. The code first reads the data from a CSV file using pandas library and selects two features, 'Annual Income (k$)' and 'Spending Score (1-100)', to form the data set. Then, it creates a scatter plot of the data points with 'Annual Income' and 'Spending Score' on the x and y axes, respectively. After that, it applies K-means clustering algorithm with 5 clusters to the data set and plots the resulting clusters along with the centroids. Finally, the code predicts the cluster label of a new data point with 'Annual Income' of 70k and 'Spending Score' of 80 and prints the predicted label.